+++ src/main/java/com/ing/Soft/services/UserService.java
@@ -1,10 +1,7 @@
 package com.ing.Soft.services;
 
-import com.ing.Soft.entities.Course;
 import com.ing.Soft.entities.User;
-import com.ing.Soft.repositories.CourseRepository;
 import com.ing.Soft.repositories.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.Optional;
 
@@ -23,7 +20,7 @@ public class UserService {
         userRepository.save(user);
     }
 
-    public Optional<User> findById(Long id){
+    public Optional<User> findById(Integer id){
         return userRepository.findById(id);
     }
